+++ app/Http/Filters/CocktailQueryFilter.php
@@ -9,6 +9,7 @@ use Kami\Cocktail\Models\Cocktail;
 use Spatie\QueryBuilder\AllowedSort;
 use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\AllowedFilter;
+use Spatie\QueryBuilder\AllowedInclude;
 use Kami\Cocktail\Repository\CocktailRepository;
 
 /**
@@ -176,8 +177,7 @@ final class CocktailQueryFilter extends QueryBuilder
                         ->orderBy('cf.updated_at', $direction);
                 }),
             ])
-            ->allowedIncludes(['glass', 'method', 'user', 'navigation', 'utensils', 'createdUser', 'updatedUser'])
-            ->with('ingredients.ingredient', 'images', 'tags', 'ratings')
+            ->allowedIncludes(['glass', 'method', 'user', 'navigation', 'utensils', 'createdUser', 'updatedUser', 'images', 'tags', 'ingredients.ingredient', 'ratings'])
             ->selectRaw('cocktails.*, COUNT(ci.cocktail_id) AS total_ingredients, COUNT(ci.ingredient_id) - COUNT(ui.ingredient_id) AS missing_ingredients')
             ->leftJoin('cocktail_ingredients AS ci', 'ci.cocktail_id', '=', 'cocktails.id')
             ->leftJoin('user_ingredients AS ui', function ($query) use ($barMembership) {